requirements.txt
fastapi
uvicorn[standard]
joblib
pandas
python-dotenv
scikit-learn
PyYAML
